missing identical hits; improved sub_n

+++ bwamem.h
@@ -33,7 +33,7 @@ typedef struct {
 typedef struct {
 	int64_t rb, re;
 	int score, qb, qe, seedcov, sub, csub; // sub: suboptimal score; csub: suboptimal inside the chain
-	double sub_n;
+	int sub_n;
 } mem_alnreg_t;
 
 typedef kvec_t(mem_chain_t)  mem_chain_v;
